Add files via upload
Include Header File: This line includes the iostream header file, which is necessary for input and output operations in C++.
Using Namespace:
This line allows the program to use standard library names (like cout and cin) without needing to prefix them with std;

Main Function: This defines the main function where the execution of the program begins. It returns an integer, which is used as the exit status of the program.

User Input Prompt: The program prints a prompt asking the user to enter a number, and then reads the input from the user into the variable a.

Conditional Logic: This block of code checks if the number stored in a is even or odd. It uses the modulus operator (%) to check the remainder when a is divided by 2. If the remainder is 0, the number is even and the program prints "The given number is EVEN". Otherwise, it prints "The given number is ODD"

Return statement : This statement signifies the successful completion of the program. The 0 indicates that the program ended without errors.

Summary: The program prompts the user to enter an integer, determines whether the number is even or odd by checking the remainder when divided by 2, and then prints the result. The use of cout and cin handles the output and input operations, respectively.